Remove document accessor from json response.
document was added as an accessor in 02f4d26d34c227e065a23c77b278fe48b901f02e.
However, when we render availability as json this puts a SolrDocument in every item in the holdings.
For records with a lot of items this blows that json response's size up to a huge amount.